Add validations to prevent null fields in Admin and Doctor classes, implement logical deletion for Admin, and add listing and restoration of deleted Admins

+++ server/src/main/java/com/example/Healthtech/models/Patient.java
@@ -14,7 +14,7 @@ public class Patient {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public int id_patient;
+    public Long id_patient;
 
     public String nombre;
